Source code synced with Eugene game development kit 8.6.3

+++ source/neongdk.h
@@ -194,10 +194,10 @@ typedef enum
    void set_video_mode();
    void get_video_mode();
    void correct_depth();
-   void set_setting(const unsigned long int width,const unsigned long int height);
-   unsigned long int get_depth() const;
-   unsigned long int get_display_width() const;
-   unsigned long int get_display_height() const;
+   void set_setting(const unsigned int width,const unsigned int height);
+   unsigned int get_depth() const;
+   unsigned int get_display_width() const;
+   unsigned int get_display_height() const;
    public:
    Display();
    ~Display();
@@ -653,7 +653,7 @@ typedef enum
   {
    private:
    void check_video_mode();
-   void set_resolution(const unsigned long int width,const unsigned long int height);
+   void set_resolution(const unsigned int width,const unsigned int height);
    void screen_setup();
    public:
    Screen();
